Refactor region division into utility function
Abstracted the logic for dividing regions into a new utility function `divideIntoRegions` in `colorUtil.ts`. This simplifies the main processing script by reducing redundancy and improving code clarity through modularization.
